client/requirements.txt
kafka-python
kazoo